+++ src/parse-gui.py
@@ -176,8 +176,15 @@ def main():
             err_count += pragmatic.parse(path, scanner, substr, prepend, control_flags)
         elif any(s in scan_match for s in parsers.pylint_keywords):
             err_count += pylint.parse(path, scanner, substr, prepend, control_flags)
+        elif any(s in scan_match for s in parsers.semgrep_keywords):
+            err_count += semgrep.parse(path, scanner, substr, prepend, control_flags)
+        elif any(s in scan_match for s in parsers.sigasi_keywords):
+            err_count += sigasi.parse(path, scanner, substr, prepend, control_flags)
         elif any(s in scan_match for s in parsers.srm_keywords):
-            err_count += srm_csv.parse(path, scanner, substr, prepend, control_flags)
+            if os.path.splitext(path)[1] == ".csv":
+                err_count += srm_csv.parse(path, scanner, substr, prepend, control_flags)
+            else:
+                err_count += srm.parse(path, scanner, substr, prepend, control_flags)
         else:
             logger.error(f"Unsupported scanner. Skipped {fpath},{scanner}")
             err_count += 1